Added the functionality to run TD distributed on Kubernetes.
Enabled to possibility of deploying more than one driver in order to
avoid having a single point of failure in the cluster. The drivers can
have one of the following roles:

    Master: is responsible for coordinating the entire execution in the
cluster (sends Heartbeat messages to periodically check that the agents
are still up and running, divides the phases from the initial
configuration into multiple tasks to be distributed between the agents,
etc).

    Candidate: it can replace the current Master in case of failure.
 Candidates are responsible for periodically sending Heartbeat messages
 to the current Master to check that it is still running properly.